--- js/screens/game-screen.js
@@ -1,3 +1,0 @@
-export default class GameScreen {
-
-}
